Moved GenericTask to common
